% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_birds.R
\name{all_birds}
\alias{all_birds}
\title{Returns dataframe of bird name information including common name, scientific name, species code, family name...}
\usage{
all_birds(key)
}
\arguments{
\item{key}{The users eBird key}
}
\value{
dataframe of all bird name data
}
\description{
Returns dataframe of bird name information including common name, scientific name, species code, family name...
}
