% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_countries.R
\name{all_countries}
\alias{all_countries}
\title{Returns dataframe of all countries}
\usage{
all_countries(key)
}
\arguments{
\item{key}{The users eBird key}
}
\value{
countries dataframe
}
\description{
Returns dataframe of all countries
}
