% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_environment.R
\name{annotation}
\alias{annotation}
\alias{weather}
\title{annotation}
\usage{
weather(x = tracks_step1, era5 = era, var = c("U10m", "V10m"), unit = "hours")
}
\arguments{
\item{unit}{the annotation unit (hours)}

\item{TRACKS}{sf object with tracks}

\item{ERA}{path/file/API ERA5 dataset with annotation variables,}

\item{VAR}{vector with ERA5 variables to be annotated}
}
\value{
returns the tracks with era5 annotation
}
\description{
annotate the trajectory with environmental data
}
\details{

}
\examples{


}
