% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_day.R
\name{bird_day}
\alias{bird_day}
\title{Returns dataframe of regions and the birds present on a given day
Can also give list of all countries if region_type is set to "country"}
\usage{
bird_day(country = "US", year, month, day, key)
}
\arguments{
\item{country}{two letter country abreviation, default is United States}

\item{year}{4 digit year written as a string}

\item{month}{month 1-12 written as a string}

\item{day}{day 1-31 written as a string}

\item{key}{The users eBird key}
}
\value{
bird_obs dataframe
}
\description{
Returns dataframe of regions and the birds present on a given day
Can also give list of all countries if region_type is set to "country"
}
