% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixing.R
\name{check_time_error}
\alias{check_time_error}
\title{check_time_error}
\usage{
check_time_error(time_inter, min_time_inter)
}
\arguments{
\item{time_inter}{time interval}

\item{min_time_inter}{minimum time interval}
}
\value{

}
\description{
check time error
}
