% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_UScode.R
\name{clean_ebirdlist}
\alias{clean_ebirdlist}
\title{Helper function to create tibble from ebird JSON parsed content}
\usage{
clean_ebirdlist(list)
}
\arguments{
\item{list}{list containing JSON parsed content returned from ebird_api}
}
\value{
A tibble of ebird data
}
\description{
Helper function to create tibble from ebird JSON parsed content
}
\author{
Simon Marks
}
