% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_filter_spatial.R
\name{create_location}
\alias{create_location}
\title{create_location}
\usage{
create_location(location, crs, verbose = T)
}
\arguments{
\item{location}{coordinate, vector of radar coordinates, or spatial point. Input coordinates in WGS84.}

\item{crs}{metric crs to perform calculations in}
}
\value{
a polygon with inclusion ring in coordinate reference system of the input
}
\description{
calculate a point geometry
}
\details{

}
