% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bias_JENS.R
\name{distb}
\alias{distb}
\title{distb}
\usage{
distb(x, track_raster, radar, distance = c(mid, mad))
}
\arguments{
\item{x}{track dataset}

\item{track_raster}{spatraster generated by track_density_rast}

\item{radar}{radar location}

\item{distance}{default mid mad}
}
\value{
a raster with counts per grid cell
}
\description{
correction for distance bias based on track raster
}
\examples{


\dontrun{

}

}
