% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{era}
\alias{era}
\title{era}
\format{
\subsection{\code{era}}{

\describe{
\item{date_time}{date and hour of the day}
\item{lat, lon}{latitude and longitude in EPSG:4326}
\item{U10m, V10m}{U and V component of era5 dataset}
}
}
}
\source{
\url{https://cds.climate.copernicus.eu/#!/search?text=ERA5&type=dataset}
}
\usage{
era
}
\description{
era5 data for the Netherlands
}
\keyword{datasets}
