% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_filter_spatial.R
\name{exclusion_location}
\alias{exclusion_location}
\alias{exclusion_geom}
\title{exclusion_location}
\usage{
exclusion_geom(x = turbines, buffer = NULL, crs = loc_epsg, verbose = F)
}
\arguments{
\item{x}{coordinate, vector of radar coordinates, or spatial point. Input coordinates in WGS84.}

\item{buffer}{proj4string to transform to for calculations in a metric system}

\item{crs}{crs to perform calculations in}

\item{y}{vector of minimum and maximum angles in degrees to exclude}
}
\value{
a polygon with inclusion donut
}
\description{
Return the inclusion ring defined for the radar location and a minimum and maximum distance
}
\details{

}
\examples{

turbines
x<- turbines
buffer <- 100 # distance in meter
crs <- 23095 # metric CRS in which calculations are performed

p <- exclusion_geom(x,buffer,crs,verbose=T)

x <- st_buffer(x$geometry,200)
p <- exclusion_geom(x,buffer,crs,verbose=T)

}
