% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixing.R
\name{first_last_removed}
\alias{first_last_removed}
\title{first_last_removed}
\usage{
first_last_removed(good_points, n)
}
\arguments{
\item{good_points}{good points of the trajectory}

\item{n}{number of points of a trajectory}
}
\value{

}
\description{
Check if the first or last point was removed
}
