% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixing.R
\name{fix_direction}
\alias{fix_direction}
\title{fix_direction}
\usage{
fix_direction(trajectory, good_points)
}
\arguments{
\item{trajectory}{trajectory geometry}

\item{good_points}{good points of the trajectory}
}
\value{

}
\description{
function to correct the direction of a trajectory
}
