% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fixing.R
\name{fix_duration}
\alias{fix_duration}
\title{fix_duration}
\usage{
fix_duration(timestamp_start, timestamp_end)
}
\arguments{
\item{timestamp_start}{start time of the trajectory}

\item{timestamp_end}{end time of the trajectory}
}
\value{

}
\description{
function to correct the duration of a trajectory
}
