% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixing.R
\name{fix_str_distance}
\alias{fix_str_distance}
\title{fix_str_distance}
\usage{
fix_str_distance(trajectory, good_points)
}
\arguments{
\item{trajectory}{trajectory geometry}

\item{good_points}{good points of the trajectory}
}
\value{

}
\description{
function to correct the displacement of a trajectory
}
