% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fixing.R
\name{fix_time_end}
\alias{fix_time_end}
\title{fix_time_end}
\usage{
fix_time_end(timestamp_start, trajectory_time)
}
\arguments{
\item{timestamp_start}{start time of the trajectory}

\item{trajectory_time}{trajectory times in seconds}
}
\value{

}
\description{
function to correct the end time corresponding to a trajectory
}
