% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixing.R
\name{fix_traj}
\alias{fix_traj}
\title{fix_traj}
\usage{
fix_traj(trajectory, good_points)
}
\arguments{
\item{trajectory}{trajectory geometry}

\item{good_points}{good points of the trajectory}
}
\value{

}
\description{
function to correct the trajectory by retaining only the good points
}
