% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fixing.R
\name{fix_vec}
\alias{fix_vec}
\title{fix_vec}
\usage{
fix_vec(vector, good_points)
}
\arguments{
\item{vector}{vector of values}

\item{good_points}{good points of the trajectory / vector}
}
\value{

}
\description{
function to correct a vector by retaining only the good points
}
