% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_day.R
\name{flock_plot}
\alias{flock_plot}
\title{Creates a leaflet map of birds spotted on a single day for a region
To be used with bird_day as this function is expecting the data from bird_day}
\usage{
flock_plot(data)
}
\arguments{
\item{data}{a dataframe prepared by bird_day}
}
\value{
a leaflet map of birds for a specific day!
}
\description{
returns dataframe of all birds spotted on a single day for a region
}
\examples{
dylan_bday <- bird_day(year = "1999", month = "1", day = "28", key = my_birdkey)
flock_plot(dylan_bday)

}
