% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_annotation_movement.R
\name{get_airspeed}
\alias{get_airspeed}
\title{get_airspeed}
\usage{
get_airspeed(groundspeed, direction, wind_u, wind_v)
}
\arguments{
\item{groundspeed}{groundspeed}

\item{direction}{direction}

\item{wind_u}{wind u / wind speed}

\item{wind_v}{wind v / wind direction}
}
\value{
airspeed in meter per second
}
\description{
calculate airspeed using u- and v-component, average groundspeed and direction
}
