% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_movement.R
\name{get_direction}
\alias{get_direction}
\title{get_direction}
\usage{
get_direction(trajectory, n)
}
\arguments{
\item{trajectory}{trajectory column}

\item{n}{number of points}
}
\description{
calculate direction using a vector of track geometries and a vector with the number of locations in a track
}
