% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_movement.R
\name{get_displacement}
\alias{get_displacement}
\title{get_displacement}
\usage{
get_displacement(trajectory, n)
}
\arguments{
\item{trajectory}{trajectory column}

\item{n}{number of points}
}
\description{
calculate displacement using a vector of track geometries and a vector with the number of locations in a track
}
