% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_filter_nonbio.R
\name{get_dot_percs}
\alias{get_dot_percs}
\title{get_dot_percs}
\usage{
get_dot_percs(x, probs = seq(0.002, 0.02, 0.002), context = NULL)
}
\arguments{
\item{x}{sf object with tracks}

\item{probs}{vector of probability bins}

\item{context}{boolean (TRUE/FALSE) to plot the tracks of each probability bin}
}
\value{
returns dot perc vals
}
\description{
visualize percentiles
}
\details{

}
\examples{


}
