% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fixing.R
\name{get_good_points}
\alias{get_good_points}
\title{get_good_points}
\usage{
get_good_points(time_inter, min_time_inter)
}
\arguments{
\item{time_inter}{time interval}

\item{min_time_inter}{minimum time interval}
}
\value{

}
\description{
return the good points
}
