% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hotspots.R
\name{get_hotspots}
\alias{get_hotspots}
\title{Returns dataframe of birding hotspots for a given region. Must enter the regions code for region, for example if you wanted hotspots
in the United States you would enter "region = \code{US}".}
\usage{
get_hotspots(region, back = 1, key)
}
\arguments{
\item{region}{The region code for the desired region}

\item{back}{How many days back to collect data, default is 1}

\item{key}{The users eBird key}
}
\value{
subregions dataframe
}
\description{
Returns dataframe of birding hotspots for a given region. Must enter the regions code for region, for example if you wanted hotspots
in the United States you would enter "region = \code{US}".
}
