% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_hotspots.R
\name{get_local_hotspots}
\alias{get_local_hotspots}
\title{Returns dataframe of local birding hotspots for a given latitude and longitude.}
\usage{
get_local_hotspots(lat, lng, dist = 25, back = 1, key)
}
\arguments{
\item{lat}{Latitude, must be to two decimal places}

\item{lng}{Longitude, must be to two decimal places}

\item{dist}{The search radius of given lat and lng in kilometers, default of 25km}

\item{back}{How many days back to collect data, default is 1}

\item{key}{The users eBird key}
}
\value{
subregions datafram
}
\description{
Returns dataframe of local birding hotspots for a given latitude and longitude.
}
