% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notable_chart.R
\name{get_notable_chart}
\alias{get_notable_chart}
\title{Creates dot plot of highest recent notable bird observation counts for a specified region within the United States}
\usage{
get_notable_chart(
  region_code,
  ebirdkey,
  obs_to_plot = 15,
  back = 14,
  hotspot = FALSE,
  includeProvisonal = FALSE,
  maxResults = 10000
)
}
\arguments{
\item{region_code}{United States subnational1 or subnational2 region code}

\item{ebirdkey}{User ebird key}

\item{obs_to_plot}{Number of observations to plot, default is 15, max is 25}

\item{back}{The number of days back to fetch observations, max is 30}

\item{hotspot}{Only fetch observations from hotspots- logical}

\item{maxResults}{Only fetch this number of observations, max is 10000}

\item{includeProvisional}{Include observations which have not yet been reviewed- logical}
}
\value{
A dotplot of highest recent notable bird observation counts
}
\description{
Creates dot plot of highest recent notable bird observation counts for a specified region within the United States. Will plot up to the top 25 most notably sighted birds for a region. Accesses data from ebird API.
}
\note{
Use pick_UScode() to determine region code for region_code argument
}
\examples{
get_notable_chart(region_code = "US-CA", ebirdkey = 'mykey', obs_to_plot = 25, back = 30)

}
\author{
Simon Marks
}
