% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notable_obs.R
\name{get_notable_obs}
\alias{get_notable_obs}
\title{Creates tibble of recent notable bird observations for supplied United States region from ebird API}
\usage{
get_notable_obs(
  region_code,
  ebirdkey,
  back = 14,
  hotspot = FALSE,
  includeProvisonal = FALSE,
  maxResults = 10000
)
}
\arguments{
\item{region_code}{United States subnational1 or subnational2 region code}

\item{ebirdkey}{User ebird key}

\item{back}{The number of days back to fetch observations}

\item{hotspot}{Only fetch observations from hotspots}

\item{maxResults}{Only fetch this number of observations}

\item{includeProvisional}{Include observations which have not yet been reviewed}
}
\value{
Tibble of recent notable bird observations ordered from most to least recent for the specified number of days back to collect records
}
\description{
Creates tibble of recent notable bird observations for a region using a United States subnational1 or subnational2 region code
}
\note{
Use pick_UScode() to determine region code for region_code argument
}
\examples{
get_notable_obs(region = "US-CA-079", ebirdkey = 'mykey', back = 15)

}
