% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regions.R
\name{get_regions}
\alias{get_regions}
\title{Returns dataframe of subregions and their codes for a parent country or a parent subregion
Can also give list of all countries if region_type is set to "country"}
\usage{
get_regions(region_type, parent_region = "world", key)
}
\arguments{
\item{region_type}{The options for this are country, subnational1, and subnational2}

\item{parent_region}{The options for this are world (the default),}

\item{key}{The users eBird key}
}
\value{
subregions datafram
}
\description{
Returns dataframe of subregions and their codes for a parent country or a parent subregion
Can also give list of all countries if region_type is set to "country"
}
