% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fixing.R
\name{get_time_inter}
\alias{get_time_inter}
\alias{get_dt}
\title{get_time_inter}
\usage{
get_dt(trajectory_time)
}
\arguments{
\item{trajectory_time}{}
}
\value{
calculate the interval in seconds between consecutive locations in a trajectory
}
\description{
time interval between consecutive points
}
\details{

}
