% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landmask}
\alias{landmask}
\title{landmask}
\format{
\subsection{\code{landmask}}{

\describe{
\item{timestamp}{UTC time}
\item{landmask}{landmask value}
}
}
}
\usage{
landmask
}
\description{
landmask exported from the luchterduinen robin database
}
\keyword{datasets}
