% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_hotspots.R
\name{local_hotspots}
\alias{local_hotspots}
\title{Returns dataframe of local birding hotspots for a given latitude and longitude.}
\usage{
local_hotspots(lat, lng, dist = 25, back = 1, key)
}
\arguments{
\item{lat}{Latitude}

\item{lng}{Longitude}

\item{dist}{The search radius of given lat and lng in kilometers, default of 25km}

\item{back}{How many days back to collect data, default is 1}

\item{key}{The users eBird key}
}
\value{
subregions datafram
}
\description{
Returns dataframe of local birding hotspots for a given latitude and longitude.
}
