% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_local_hotspots.R
\name{map_local_hotspots}
\alias{map_local_hotspots}
\title{Returns map of local hotspots. Labels each hotspots title. Popup will include name, lat/lng, latest obeservation
and the total number of species observed}
\usage{
map_local_hotspots(
  latitude,
  longitude,
  back = 1,
  dist = 25,
  iconwidth = 30,
  iconheight = 30,
  key
)
}
\arguments{
\item{back}{How many days back to collect data, default is 1}

\item{dist}{The search radius of given lat and lng in kilometers, default of 25km}

\item{iconwidth}{Width of map icon, default is 30}

\item{iconheight}{Height of map icon, default is 30}

\item{key}{The users eBird key}

\item{lat}{Latitude, must be to two decimal places}

\item{lng}{Longitude, must be to two decimal places}
}
\value{
countries dataframe
}
\description{
Returns map of local hotspots. Labels each hotspots title. Popup will include name, lat/lng, latest obeservation
and the total number of species observed
}
