% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_movement.R
\name{movement}
\alias{movement}
\title{movement}
\usage{
movement(
  TRACKS,
  VAR = c("n", "duration", "length", "groundspeed", "displacement", "direction", "dot"),
  crs = loc_epsg
)
}
\arguments{
\item{TRACKS}{sf object with tracks}

\item{VAR}{vector with movement metrics to add to tracks}

\item{crs}{}
}
\value{
returns annotated trajectories with movement information
}
\description{
annotate the trajectory with movement data
}
\details{

}
\examples{


}
