% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_detecting_negative_observation_bias.R
\name{nob_gam}
\alias{nob_gam}
\title{nob_gam}
\usage{
nob_gam(aggr_stats)
}
\arguments{
\item{aggr_stats}{track data with minimum the geometry, timestamp_start and timestamp_end}
}
\value{
returns the tracks with era5 annotation
}
\description{
negative observation bias based on mask
}
\details{

}
\examples{


}
