% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_UScode.R
\name{pick_UScode}
\alias{pick_UScode}
\title{Determines subnational1 and subnational2 region codes for United States from ebird API}
\usage{
pick_UScode(state, county = NULL, ebirdkey)
}
\arguments{
\item{state}{String of full US state name (i.e. "California")}

\item{county}{String of full county name (i.e "San Luis Obispo"), only required if subnational2 region code is desired}

\item{ebirdkey}{User ebird key}
}
\value{
Tibble containing region name and code
}
\description{
Determines subnational1 and subnational2 region codes for United States from ebird API, useful for deciding region arguments in some Birdr functions
}
\note{
Some counties will result in a return of multiple subnational2 ebird region codes because there are different counties with the same name.
}
\examples{
pick_UScode(state = "California", county = "San Luis Obispo", ebirdkey = 'mykey'))

}
