% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_detecting_spatial_bias.R
\name{rasterize_basic}
\alias{rasterize_basic}
\title{rasterize_basic}
\usage{
rasterize_basic(
  data = NULL,
  resolution = 100,
  crs = NULL,
  fun = "weighted",
  e = NULL
)
}
\arguments{
\item{data}{spatvector object (points)}

\item{resolution}{resolution}

\item{fun}{summary function}

\item{e}{minimum x value of extent in lonlat}

\item{aoi}{area of inclusion}
}
\value{
a raster with counts per grid cell
}
\description{
count number of points per grid cell
}
\examples{


\dontrun{
r <- grid_count(data, resolution=500)
}

}
