% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_detecting_spatial_bias.R
\name{rasterize_plot}
\alias{rasterize_plot}
\title{rasterize_plot}
\usage{
rasterize_plot(x, aoi = NULL)
}
\arguments{
\item{x}{spatraster object}

\item{aoi}{area of inclusion}
}
\value{
a plot of the raster and the area of inclusion
}
\description{
plot raster
}
\examples{


\dontrun{

}

}
