% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bias_JOH.R
\name{rasterizer}
\alias{rasterizer}
\title{rasterizer}
\usage{
rasterizer(
  data = NULL,
  resolution = 5000,
  var = NULL,
  fun = "length",
  e = NULL
)
}
\arguments{
\item{data}{spatvector object (points)}

\item{resolution}{resolution}

\item{var}{variable to loop over}

\item{e}{minimum x value of extent}
}
\value{
a raster or list of rasters with counts per grid cell
}
\description{
count number of points per grid cell
}
\examples{


\dontrun{
r <- grid_count(data, resolution=500)
}

}
