% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bias_JENS.R
\name{rastgam}
\alias{rastgam}
\title{rastgam}
\usage{
rastgam(track_raster, radar, distance = c(mid, mad))
}
\arguments{
\item{track_raster}{spatraster generated by track_density_rast}

\item{radar}{radar location}

\item{distance}{default mid mad}
}
\value{
a raster with counts per grid cell
}
\description{
count number of points per grid cell
}
\examples{


\dontrun{
r <- rastgam(data, resolution=500)
}

}
