% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spatial.R
\name{roi}
\alias{roi}
\title{roi}
\usage{
roi(
  location = NULL,
  distance = NULL,
  excl_angles = NULL,
  excl_geom = NULL,
  excl_buffer = NULL,
  crs = loc_epsg,
  verbose = T
)
}
\arguments{
\item{location}{coordinate, vector of radar coordinates, or spatial point. Input coordinates in WGS84.}

\item{crs}{crs to perform calculations in. If different from location coordinate, the location coordinate will reprojected}

\item{inclusion_ring}{vector of minimum and maximum distance in meters to include}

\item{exclusion_angle}{vector, list or data.frame of angle-pairs to exclude from analysis}

\item{exclusion_location}{sf-object with locations which can reliably be excluded}

\item{exclusion_location_buffer}{buffer in unit of crs?}
}
\value{
a polygon with inclusion zones
}
\description{
Return the area of interest. Three filters (inclusion_donut, exclusion_angle, exclusion_location) can be specified.
}
\details{
With the argument exclusion_angle for which we expect structural interference.
Here we have two types of interference:
\enumerate{
\item The turbine the radar is situated on blocks one side of the radar window
\item The area overlapping with the vertical radar beam, which mostly has mixed
tracks we did not include in this study. Therefore this area is not relevant here
We remove the blocked corners by creating a triangle using the radar coordinates,
the corner angles, and  maximum distance.
}
}
\examples{

\dontrun{
# inclusion ring

# location as vector
location <- c(4.185345, 52.42783)
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095)

# location as geom
# 4326
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095)
# 3035
location <- st_sfc(st_point(c(3925960, 3273397)),crs=3035)
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095)
## distance

# one distance value: no exclusion of area near radar
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)
area_of_inclusion <-  roi(location=location,
                         distance=2500,
                         crs=23095)
# two distance values: no exclusion of area near radar
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095)

## exclusion angles
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)

# exclusion angles can be a list, vector or matrix

# list
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=list(c(287,30),c(115,135)))

# vector
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135))

# matrix
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=matrix(c(287,30,115,135),nrow=2,ncol=2,byrow=T))

# do not exclude areas near the radar
area_of_inclusion <-  roi(location=location,
                         distance=c(2500),
                         crs=23095,
                         excl_angles=matrix(c(287,30,115,135),nrow=2,ncol=2,byrow=T))

## exclusion areas

# without buffer points will not be excluded
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)

area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=turbines)

# buffer outside the function
polygons <- st_buffer(st_transform(turbines$geometry,23095),100)
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=polygons)
# buffer within the function
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=turbines,
                         excl_buffer=100
)

# vary the buffer distance based on distance from radar
dist <- st_distance(location,turbines$geometry)
weight <- as.vector(dist/max(dist) * 100)

area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=turbines,
                         excl_buffer=weight
)

# apply a buffer around multiple locations by using lists
area_of_inclusion <-  roi(location=location,
                         distance=2500,
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=list(turbines,location),
                         excl_buffer=list(100,1000)
)

# apply a buffer around multiple locations by using lists and varying buffer weight within a single dataset
dist <- st_distance(location,turbines$geometry)
weight <- as.vector(dist/max(dist) * 100)
area_of_inclusion <-  roi(location=location,
                         distance=2500,
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=list(turbines,location),
                         excl_buffer=list(weight,1000)
)
}
}
