% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_detecting_spatial_bias.R
\name{track_density_rast}
\alias{track_density_rast}
\title{track_density_rast}
\usage{
track_density_rast(
  data = trajectories,
  crs = loc_proj,
  e = NULL,
  resolution = 100,
  aoi = NULL
)
}
\arguments{
\item{data}{spatvector object (points)}

\item{crs}{summary function}

\item{e}{minimum x value of extent in lonlat}

\item{resolution}{resolution}

\item{aoi}{aoi}
}
\value{
a raster with counts per grid cell
}
\description{
count number of points or trajectories per grid cell
}
\examples{


\dontrun{
r <- track_density_rast(data, resolution=500)
}

}
