% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracks}
\alias{tracks}
\title{tracks_all}
\format{
\subsection{\code{tracks}}{

\describe{
\item{id}{unique_identifier}
\item{timestamp_start}{UTC timestamp corresponding to the first location of a bird-track}
\item{timestamp_end}{UTC timestamp corresponding to the last location of a bird-track}
\item{trajectory_time}{offset seconds from timestamp_start for each location in a bird-track}
\item{trajectory}{linestring geometry corresponding to a bird-track}
}
}
}
\usage{
tracks
}
\description{
trajectories
}
\keyword{datasets}
