% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turbines}
\alias{turbines}
\title{turbines}
\format{
\subsection{\code{turbines}}{

\describe{
\item{OBJECTID}{unique identifier}
\item{NAAM}{windpark name}
\item{geometry}{point geometry corresponding to a turbine location}
}
}
}
\usage{
turbines
}
\description{
turbines of all windparks from the Netherlands
}
\keyword{datasets}
