% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_annotation_environment.R
\name{annotation}
\alias{annotation}
\alias{weather}
\title{annotation}
\usage{
weather(
  x = tracks_step1,
  era5 = era,
  variables = c("U10m", "V10m"),
  unit = "hours"
)
}
\arguments{
\item{x}{sf object with tracks}

\item{era5}{path/file/API ERA5 dataset with annotation variables,}

\item{variables}{vector with ERA5 variables to be annotated}

\item{unit}{the annotation unit (hours)}
}
\value{
returns the tracks with era5 annotation
}
\description{
annotate the trajectory with environmental data
}
\details{

}
\examples{


}
