% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_day.R
\name{bird_day}
\alias{bird_day}
\title{Creates a dataframe of birds spotted on a single day for a region (default United States)
For specific states or counties use pick_UScode to generate the country ID.}
\usage{
bird_day(country = "US", year, month, day, key)
}
\arguments{
\item{country}{string of country letter abreviation, default is United States}

\item{year}{string of 4 digit year}

\item{month}{string of month 1-12}

\item{day}{string of day 1-31}

\item{key}{The users eBird key}
}
\value{
bird_obs dataframe
}
\description{
returns dataframe of all birds spotted on a single day for a region
}
\examples{
dylan_bday <- bird_day(year = "1999", month = "1", day = "28", key = my_birdkey)

}
