% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bias_JENS.R
\name{crosses_aoi}
\alias{crosses_aoi}
\title{crosses_aoi}
\usage{
crosses_aoi(trajectory, raster)
}
\arguments{
\item{trajectory}{trajectory geometry}

\item{raster}{raster}
}
\value{
a raster with counts per grid cell
}
\description{
count number of points per grid cell
}
\examples{


\dontrun{
r <- rastgam(data, resolution=500)
}

}
