% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_detecting_spatial_bias.R
\name{distb}
\alias{distb}
\title{distb}
\usage{
distb(x, y, aoi, radar, distance = c(mid, mad))
}
\arguments{
\item{x}{track dataset}

\item{radar}{radar location}

\item{distance}{default mid mad}

\item{track_raster}{spatraster generated by track_density_rast}
}
\value{
a raster with counts per grid cell
}
\description{
correction for distance bias based on track raster
}
\examples{


\dontrun{
x <- tracks[sample(1:nrow(tracks),10000),]
y <- track_density_rast(tracks$trajectory,resolution=100)
radar <- sf::st_sfc(sf::st_point(c(4.185345,  52.427827)),crs=4326)
distance <- c(1000,2500)
distb(x,y, radar, distance,aoi=NULL)

}

}
