% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fixing.R
\name{fix_geom}
\alias{fix_geom}
\title{fix_geom}
\usage{
fix_geom(x, dt_min)
}
\arguments{
\item{x}{tracks}

\item{dt_min}{minimum time interval}
}
\value{
track table including only the tracks with erroneous points
}
\description{
correct geometries
}
\details{
Erroneous points exist within some tracks, identifiable by a very short time-interval
between those points and the next/previous point. Identify these intervals and remove
both points involved with that interval, as there is no way to tell which of these two
is faulty.
}
