% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nonbio.R
\name{get_dot_percs}
\alias{get_dot_percs}
\title{get_dot_percs}
\usage{
get_dot_percs(x, probs = seq(0.002, 0.02, 0.002), context = NULL)
}
\arguments{
\item{probs}{vector with ERA5 variables to be annotated}

\item{displacement}{sf object with tracks}

\item{duration}{path/file/API ERA5 dataset with annotation variables,}

\item{plot}{boolean (TRUE/FALSE) to plot the}
}
\value{
returns dot perc vals
}
\description{
visualize percentiles
}
\details{

}
\examples{


}
