% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_annotation_movement.R
\name{movement}
\alias{movement}
\title{movement}
\usage{
movement(
  x,
  metrics = c("n", "duration", "length", "groundspeed", "displacement", "direction",
    "dot"),
  crs = loc_epsg
)
}
\arguments{
\item{x}{sf object with tracks}

\item{metrics}{vector with movement metrics to add to tracks}

\item{crs}{}
}
\value{
returns annotated trajectories with movement information
}
\description{
annotate the trajectory with movement data
}
\details{

}
\examples{


}
