% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_detecting_spatial_bias.R
\name{rasterizer}
\alias{rasterizer}
\title{rasterizer}
\usage{
rasterizer(
  data = NULL,
  resolution = 5000,
  var = NULL,
  crs = NULL,
  fun = "weighted",
  e = NULL,
  aoi = NULL,
  verbose = T
)
}
\arguments{
\item{data}{spatvector object (points)}

\item{resolution}{resolution}

\item{var}{variable to loop over}

\item{crs}{coordinate reference system}

\item{e}{minimum x value of extent}

\item{aoi}{area of inclusion}

\item{verbose}{boolean (TRUE/FALSE), print messages and plot results when TRUE}
}
\value{
a raster or list of rasters with counts per grid cell
}
\description{
count number of points per grid cell
}
\examples{


\dontrun{
r <- grid_count(data, resolution=500)
}

}
