% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_detecting_spatial_bias.R
\name{rastgam}
\alias{rastgam}
\title{rastgam}
\usage{
rastgam(x, radar, distance = c(mid, mad))
}
\arguments{
\item{x}{spatraster generated by track_density_rast}

\item{radar}{radar location}

\item{distance}{default mid mad}
}
\value{
a raster with counts per grid cell
}
\description{
count number of points per grid cell
}
\examples{


\dontrun{

}

}
